+++ lib/src/common/utils.spec.ts
@@ -1,7 +1,12 @@
 import inspector from 'inspector';
 import { Pool } from 'pg';
 import { getDisabledLogger } from './logger';
-import { executeTransaction, sleep } from './utils';
+import {
+  IsolationLevel,
+  awaitWithTimeout,
+  executeTransaction,
+  sleep,
+} from './utils';
 
 const isDebugMode = (): boolean => inspector.url() !== undefined;
 if (isDebugMode()) {
@@ -24,6 +29,26 @@ describe('Utils Unit Tests', () => {
       // Assert
       const endTime = Date.now();
       expect(endTime - startTime).toBeGreaterThanOrEqual(sleepTime);
+    });
+  });
+
+  describe('awaitWithTimeout', () => {
+    it('should resolve the promise before timeout', async () => {
+      const expectedResult = 'expected result';
+      const fastPromise = () => Promise.resolve(expectedResult);
+
+      await expect(awaitWithTimeout(fastPromise, 100)).resolves.toBe(
+        expectedResult,
+      );
+    });
+
+    it('should reject with a timeout error when the promise takes too long', async () => {
+      const slowPromise = async () => sleep(400);
+      const failureMessage = 'Promise timed out!';
+
+      await expect(
+        awaitWithTimeout(slowPromise, 200, failureMessage),
+      ).rejects.toThrow(failureMessage);
     });
   });
 
@@ -50,16 +75,111 @@ describe('Utils Unit Tests', () => {
 
     it('should open a transaction and execute the callback', async () => {
       // Act
-      await executeTransaction(pool, callback, logger);
+      await executeTransaction(
+        pool,
+        callback,
+        IsolationLevel.Serializable,
+        logger,
+      );
 
       // Assert
       expect(pool.connect).toHaveBeenCalled();
       expect(callback).toHaveBeenCalled();
     });
 
+    it('should open the transaction for isolation level "read committed" correctly', async () => {
+      // Arrange
+      const client = await pool.connect();
+
+      // Act
+      await executeTransaction(
+        pool,
+        callback,
+        IsolationLevel.ReadCommitted,
+        logger,
+      );
+
+      // Assert
+      expect(client.query).toHaveBeenCalledWith(
+        'START TRANSACTION ISOLATION LEVEL READ COMMITTED',
+      );
+      expect(client.release).toHaveBeenCalled();
+    });
+
+    it('should open the transaction for isolation level "repeatable read" correctly', async () => {
+      // Arrange
+      const client = await pool.connect();
+
+      // Act
+      await executeTransaction(
+        pool,
+        callback,
+        IsolationLevel.RepeatableRead,
+        logger,
+      );
+
+      // Assert
+      expect(client.query).toHaveBeenCalledWith(
+        'START TRANSACTION ISOLATION LEVEL REPEATABLE READ',
+      );
+      expect(client.release).toHaveBeenCalled();
+    });
+
+    it('should open the transaction for isolation level "serializable" correctly', async () => {
+      // Arrange
+      const client = await pool.connect();
+
+      // Act
+      await executeTransaction(
+        pool,
+        callback,
+        IsolationLevel.Serializable,
+        logger,
+      );
+
+      // Assert
+      expect(client.query).toHaveBeenCalledWith(
+        'START TRANSACTION ISOLATION LEVEL SERIALIZABLE',
+      );
+      expect(client.release).toHaveBeenCalled();
+    });
+
+    it('should open the transaction for a "not defined" isolation level correctly', async () => {
+      // Arrange
+      const client = await pool.connect();
+
+      // Act
+      await executeTransaction(pool, callback);
+
+      // Assert
+      expect(client.query).toHaveBeenCalledWith('BEGIN');
+      expect(client.release).toHaveBeenCalled();
+    });
+
+    it('should open a default transaction for an invalid isolation level', async () => {
+      // Arrange
+      const client = await pool.connect();
+
+      // Act
+      await executeTransaction(
+        pool,
+        callback,
+        'DROP TABLE outbox; --' as IsolationLevel,
+      );
+
+      // Assert
+      expect(client.query).toHaveBeenCalledWith('BEGIN');
+      expect(client.release).toHaveBeenCalled();
+    });
+
     it('should return the result of the callback', async () => {
       // Act
-      const result = await executeTransaction(pool, callback, logger);
+      const result = await executeTransaction(
+        pool,
+        callback,
+        IsolationLevel.Serializable,
+        logger,
+      );
 
       // Assert
       expect(result).toBe('SELECT 1;');
@@ -70,7 +190,12 @@ describe('Utils Unit Tests', () => {
       const client = await pool.connect();
 
       // Act
-      await executeTransaction(pool, callback, logger);
+      await executeTransaction(
+        pool,
+        callback,
+        IsolationLevel.Serializable,
+        logger,
+      );
 
       // Assert
       expect(client.query).toHaveBeenCalledWith('COMMIT');
@@ -84,7 +209,7 @@ describe('Utils Unit Tests', () => {
 
       // Act + Assert
       await expect(
-        executeTransaction(pool, callback, logger),
+        executeTransaction(pool, callback, IsolationLevel.Serializable, logger),
       ).rejects.toThrow();
       expect(client.query).toHaveBeenCalledWith('ROLLBACK');
       expect(client.release).toHaveBeenCalled();